requirements.txt
SpeechRecognition
pyngrok 
streamlit 
streamlit-audiorecorder 
langchain 
google-search-results 
soundfile 
openai 
pydub
setuptools==64.0.2 
setuptools-rust